dairyTrack-api/MilkProduction/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
python-dotenv==0.20.0
Werkzeug==2.2.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.2
Flask-JWT-Extended==4.4.4
pandas==1.5.3
openpyxl==3.0.10
fpdf2==2.5.6
apscheduler==3.10.1
pytz ==2023.3